// resources/js/Pages/Admin/Settings/types.ts
export interface AuthSettings {
    enable_registration: boolean;
    require_email_verification: boolean;
    recaptcha_enabled: boolean;
    recaptcha_site_key: string;
    recaptcha_has_secret: boolean;
    google_login_enabled: boolean;
    google_client_id: string;
    google_has_secret: boolean;
    facebook_login_enabled: boolean;
    facebook_client_id: string;
    facebook_has_secret: boolean;
    github_login_enabled: boolean;
    github_client_id: string;
    github_has_secret: boolean;
    session_timeout: number;
    password_min_length: number;
}